requirements.txt
requests
pyyaml
humanize
